Allow custom properties marked as public to be returned as authorization/logout/token response parameters
